requirements.txt
pytest==7.4.0
azure-devops==7.1.0b3
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
pandas==2.1.0
